customize sample page
Here we create two sub navigation bar, they are add pc and view pc.
Though we don't create the database for add pc but later we will create
it. Btw, here the primary user  add pc information like the name of the
pc and it's configuration that the customer will use it. Here , I just
enter the data but it will no show my database cause I have not create
it